road-sign-detection/requirements.txt
Flask~=3.1.1

tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0
pandas==1.5.3
scikit-learn==1.2.2